refactor: share case law search policy SQL
